pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wonderwall"
version = "0.1.0"
description = "Wonderwall — Kirk-to-LLM projection adapter and inference pipeline"
requires-python = ">=3.11"
authors = [
    { name = "Kavara", email = "hello@kavara.ai" },
]
license = { text = "Apache-2.0" }
readme = "README.md"
dependencies = [
    "torch>=2.3",
    "numpy>=1.26",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "click>=8.1",
]

[project.optional-dependencies]
llm = [
    "transformers>=4.45",
    "accelerate>=0.34",
    "tokenizers>=0.20",
]
scotty = [
    "httpx>=0.27",
]
kafka = [
    "confluent-kafka>=2.4",
    "python-json-logger>=2.0",
]
serve = [
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "prometheus-client>=0.21",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.6",
    "httpx>=0.27",  # used by fastapi.testclient
]
all = [
    "wonderwall[llm,scotty,kafka,serve]",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["wonderwall*"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-q --strict-markers"

[tool.ruff]
line-length = 100
target-version = "py311"
